chore: refactor signup page logic
