+++ autogpt_platform/frontend/src/app/(platform)/signup/page.tsx
@@ -1,6 +1,4 @@
 "use client";
-import { signup } from "./actions";
-import { providerLogin } from "../login/actions";
 import {
   Form,
   FormControl,
@@ -10,15 +8,9 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-import { useForm } from "react-hook-form";
 import { Input } from "@/components/ui/input";
-import type { z } from "zod";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useCallback, useState } from "react";
-import { useRouter } from "next/navigation";
 import Link from "next/link";
 import { Checkbox } from "@/components/ui/checkbox";
-import useSupabase from "@/lib/supabase/useSupabase";
 import LoadingBox from "@/components/ui/loading";
 import {
   AuthCard,
@@ -30,88 +22,28 @@ import {
   Turnstile,
 } from "@/components/auth";
 import AuthFeedback from "@/components/auth/AuthFeedback";
-import { signupFormSchema, LoginProvider } from "@/types/auth";
 import { getBehaveAs } from "@/lib/utils";
-import { useTurnstile } from "@/hooks/useTurnstile";
+import { useSignupPage } from "./useSignupPage";
 
 export default function SignupPage() {
-  const { supabase, user, isUserLoading } = useSupabase();
-  const [feedback, setFeedback] = useState<string | null>(null);
-  const router = useRouter();
-  const [isLoading, setIsLoading] = useState(false);
-  const [isGoogleLoading, setIsGoogleLoading] = useState(false);
-  //TODO: Remove after closed beta
+  const {
+    form,
+    feedback,
+    turnstile,
+    isLoggedIn,
+    isLoading,
+    isUserLoading,
+    isGoogleLoading,
+    isSupabaseAvailable,
+    handleSignup,
+    handleProviderSignup,
+  } = useSignupPage();
 
-  const turnstile = useTurnstile({
-    action: "signup",
-    autoVerify: false,
-    resetOnError: true,
-  });
-
-  const form = useForm<z.infer<typeof signupFormSchema>>({
-    resolver: zodResolver(signupFormSchema),
-    defaultValues: {
-      email: "",
-      password: "",
-      confirmPassword: "",
-      agreeToTerms: false,
-    },
-  });
-
-  const onProviderLogin = useCallback(async (provider: LoginProvider) => {
-    setIsGoogleLoading(true);
-    const error = await providerLogin(provider);
-    setIsGoogleLoading(false);
-    if (error) {
-      setFeedback(error);
-      return;
-    }
-    setFeedback(null);
-  }, []);
-
-  const onSignup = useCallback(
-    async (data: z.infer<typeof signupFormSchema>) => {
-      setIsLoading(true);
-
-      if (!(await form.trigger())) {
-        setIsLoading(false);
-        return;
-      }
-
-      if (!turnstile.verified) {
-        setFeedback("Please complete the CAPTCHA challenge.");
-        setIsLoading(false);
-        return;
-      }
-
-      const error = await signup(data, turnstile.token as string);
-      setIsLoading(false);
-      if (error) {
-        if (error === "user_already_exists") {
-          setFeedback("User with this email already exists");
-          turnstile.reset();
-          return;
-        } else {
-          setFeedback(error);
-          turnstile.reset();
-        }
-        return;
-      }
-      setFeedback(null);
-    },
-    [form, turnstile],
-  );
-
-  if (user) {
-    console.debug("User exists, redirecting to /");
-    router.push("/");
-  }
-
-  if (isUserLoading || user) {
+  if (isUserLoading || isLoggedIn) {
     return <LoadingBox className="h-[80vh]" />;
   }
 
-  if (!supabase) {
+  if (!isSupabaseAvailable) {
     return (
       <div>
         User accounts are disabled because Supabase client is unavailable
@@ -126,7 +58,7 @@ export default function SignupPage() {
       {/* Google OAuth Button */}
       <div className="mb-6">
         <GoogleOAuthButton
-          onClick={() => onProviderLogin("google")}
+          onClick={() => handleProviderSignup("google")}
           isLoading={isGoogleLoading}
           disabled={isLoading}
         />
@@ -140,7 +72,7 @@ export default function SignupPage() {
       </div>
 
       <Form {...form}>
-        <form onSubmit={form.handleSubmit(onSignup)}>
+        <form onSubmit={form.handleSubmit(handleSignup)}>
           <FormField
             control={form.control}
             name="email"
@@ -201,7 +133,7 @@ export default function SignupPage() {
           />
 
           <AuthButton
-            onClick={() => onSignup(form.getValues())}
+            onClick={() => handleSignup(form.getValues())}
             isLoading={isLoading}
             type="submit"
           >
